method save,getall,getid and update created with sucess

+++ src/main/java/com/cadastro/pessoas/dto/PersonRecordDto.java
@@ -3,5 +3,7 @@ package com.cadastro.pessoas.dto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record PersonRecordDto(@NotBlank String name, @NotNull Integer age, @NotNull Double saldo, @NotBlank String cpf) {
+import java.math.BigDecimal;
+
+public record PersonRecordDto(@NotBlank String name, @NotNull Integer age, @NotNull BigDecimal saldo, @NotBlank String cpf) {
 }
